refactor: view 레이어 및 출력 메세지 상수 각 도메인에 맞게 분리
