+++ src/main/java/store/payment/view/PaymentOutputView.java
@@ -1,26 +1,13 @@
-package store.view;
+package store.payment.view;
 
 import java.util.List;
 import store.payment.dto.BenefitProductReceipt;
 import store.payment.dto.PaymentReceipt;
 import store.payment.dto.PurchaseProductReceipt;
-import store.promotion.dto.NonPromotionProduct;
-import store.product.dto.ProductStatus;
-import store.promotion.dto.PromotionProduct;
 
-public interface OutputView {
-
-    void printIntro();
-
-    void printProductStatus(final List<ProductStatus> productStatuses);
-
-    void printAskPurchaseProduct();
-
-    void printGuideNonPromotionProduct(final NonPromotionProduct nonPromotionProduct);
+public interface PaymentOutputView {
 
     void printAskMemberShipDiscount();
-
-    void printGuidePromotionProduct(final PromotionProduct promotionProduct);
 
     void printPurchaseProductReceipts(final List<PurchaseProductReceipt> purchaseProductReceipts);
 
@@ -29,4 +16,5 @@ public interface OutputView {
     void printPaymentReceipt(final PaymentReceipt paymentReceipt);
 
     void printAskRetryPurchase();
+
 }
