+++ src/main/java/store/product/view/ProductOutputView.java
@@ -1,0 +1,13 @@
+package store.product.view;
+
+import java.util.List;
+import store.product.dto.ProductStatus;
+
+public interface ProductOutputView {
+
+    void printIntro();
+
+    void printProductStatus(final List<ProductStatus> productStatuses);
+
+    void printAskPurchaseProduct();
+}
